+++ src/main/java/com/ray/pomin/order/controller/dto/OrderRequest.java
@@ -1,4 +1,4 @@
-package com.ray.pomin.customer.controller.dto;
+package com.ray.pomin.order.controller.dto;
 
 public record OrderRequest(String receiptNumber, String request, int cookingMinute, long storeId) {
 }
